Added some basic commands and more config options available.
Now worlds can be specified not to be loaded on startup.
Now worlds are always saved asynchronously unless the server has already stopped.
Current commands are /smw <help|version|load|unload|goto>.
If you try to load a world that was serialized with a newer version than the one SMW supports, a NewerFormatException will be thrown.

+++ slimeworldmanager-api/src/main/java/com/grinderwolf/smw/api/utils/SlimeFormat.java
@@ -3,5 +3,5 @@ package com.grinderwolf.smw.api.utils;
 public class SlimeFormat {
 
     public static final byte[] SLIME_HEADER = new byte[] { -79, 11 };
-    public static final int SLIME_VERSION = 3;
+    public static final byte SLIME_VERSION = 3;
 }
